+++ src/main/java/com/scholarly/entity/Role.java
@@ -17,7 +17,7 @@ public class Role {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private short id;
 
     @Column(nullable = false, unique = true)
     private String name;
