+++ src/main/resources/db/migration/V1__Create_roles_and_users_tables.sql
@@ -1,16 +1,16 @@
 -- roles relation
 CREATE TABLE roles (
-    id SERIAL PRIMARY KEY,
+    id SMALLSERIAL PRIMARY KEY,
     name VARCHAR(255) NOT NULL,
     description TEXT
 );
 
 -- users relation
 CREATE TABLE users (
-    id SERIAL PRIMARY KEY,
+    id BIGSERIAL PRIMARY KEY,
     username VARCHAR(255) NOT NULL UNIQUE,
     password VARCHAR(255) NOT NULL,
-    role_id INT,
+    role_id SMALLINT,
     CONSTRAINT fk_role
       FOREIGN KEY(role_id) REFERENCES roles(id)
 );
