+++ src/main/resources/db/migration/V3__Topic_membership.sql
@@ -1,7 +1,7 @@
 -- memberships table
 CREATE TABLE memberships (
-    profile_id INT NOT NULL,
-    topic_id INT NOT NULL,
+    profile_id BIGINT NOT NULL,
+    topic_id BIGINT NOT NULL,
     power SMALLINT,
     CONSTRAINT fk_profile FOREIGN KEY(profile_id) REFERENCES profiles(id),
     CONSTRAINT fk_topic_membership FOREIGN KEY(topic_id) REFERENCES topics(id),
@@ -9,8 +9,8 @@ CREATE TABLE memberships (
 );
 
 CREATE TABLE join_requests (
-    profile_id INT NOT NULL,
-    topic_id INT NOT NULL,
+    profile_id BIGINT NOT NULL,
+    topic_id BIGINT NOT NULL,
     CONSTRAINT fk_profile_requests FOREIGN KEY(profile_id) REFERENCES profiles(id),
     CONSTRAINT fk_topic_requests FOREIGN KEY(topic_id) REFERENCES topics(id),
     CONSTRAINT unique_profile_topic_requests UNIQUE (profile_id, topic_id)
